Fixed clippy warnings, split out zset and checked int
