+++ src/algebra/mod.rs
@@ -1,21 +1,14 @@
-/*
-MIT License
-SPDX-License-Identifier: MIT
-
-Copyright (c) 2021 VMware, Inc
-*/
-
 //! This module contains declarations of abstract algebraic concepts:
 //! monoids, groups, rings, etc.
+
+mod checked_int;
 
 pub mod finite_map;
 pub mod zset;
 
-use num::{traits::CheckedNeg, CheckedAdd, CheckedMul, CheckedSub};
-use std::{
-    fmt::{Display, Error, Formatter},
-    ops::{Add, AddAssign, Mul, Neg},
-};
+pub use checked_int::CheckedInt;
+
+use std::ops::{Add, AddAssign, Mul, Neg};
 
 /// A trait for types that have a zero value.
 /// This is simlar to the standard Zero trait, but that
@@ -199,9 +192,6 @@ where
     }
 }
 
-/////////////////////////////////////////////////////
-/// Ring on i64 values
-
 #[cfg(test)]
 mod integer_ring_tests {
     use super::*;
@@ -224,194 +214,5 @@ mod integer_ring_tests {
         assert_eq!(2, two);
         assert_eq!(-2, two.neg_by_ref());
         assert_eq!(-4, two.mul_by_ref(&two.neg_by_ref()));
-    }
-}
-
-////////////////////////////////////////////////////////
-/// Ring on numeric values that panics on overflow
-/// Computes exactly like any signed numeric value, but panics on overflow
-#[derive(PartialEq, Debug, Eq, Clone)]
-pub struct CheckedInt<T> {
-    value: T,
-}
-
-impl<T> CheckedInt<T> {
-    fn new(value: T) -> Self {
-        Self { value }
-    }
-}
-
-impl<T> Add for CheckedInt<T>
-where
-    T: CheckedAdd,
-{
-    type Output = Self;
-
-    fn add(self, other: Self) -> Self {
-        // intentional panic on overflow
-        Self {
-            value: self.value.checked_add(&other.value).expect("overflow"),
-        }
-    }
-}
-
-impl<T> AddByRef for CheckedInt<T>
-where
-    T: CheckedAdd,
-{
-    fn add_by_ref(&self, other: &Self) -> Self {
-        // intentional panic on overflow
-        Self {
-            value: self.value.checked_add(&other.value).expect("overflow"),
-        }
-    }
-}
-
-impl<T> AddAssign for CheckedInt<T>
-where
-    T: CheckedAdd,
-{
-    fn add_assign(&mut self, other: Self) {
-        self.value = self.value.checked_add(&other.value).expect("overflow")
-    }
-}
-
-impl<T> AddAssignByRef for CheckedInt<T>
-where
-    T: CheckedAdd,
-{
-    fn add_assign_by_ref(&mut self, other: &Self) {
-        self.value = self.value.checked_add(&other.value).expect("overflow")
-    }
-}
-
-impl<T> MulByRef for CheckedInt<T>
-where
-    T: CheckedMul,
-{
-    fn mul_by_ref(&self, rhs: &Self) -> Self {
-        // intentional panic on overflow
-        Self {
-            value: self.value.checked_mul(&rhs.value).expect("overflow"),
-        }
-    }
-}
-
-impl<T> NegByRef for CheckedInt<T>
-where
-    T: CheckedNeg,
-{
-    fn neg_by_ref(&self) -> Self {
-        Self {
-            // intentional panic on overflow
-            value: self.value.checked_neg().expect("overflow"),
-        }
-    }
-}
-
-impl<T> Neg for CheckedInt<T>
-where
-    T: CheckedNeg,
-{
-    type Output = Self;
-
-    fn neg(self) -> Self {
-        Self {
-            // intentional panic on overflow
-            value: self.value.checked_neg().expect("overflow"),
-        }
-    }
-}
-
-impl<T> HasZero for CheckedInt<T>
-where
-    T: ::num::traits::Zero + CheckedAdd,
-{
-    fn is_zero(&self) -> bool {
-        T::is_zero(&self.value)
-    }
-    fn zero() -> Self {
-        CheckedInt::new(T::zero())
-    }
-}
-
-impl<T> HasOne for CheckedInt<T>
-where
-    T: ::num::traits::One + CheckedMul,
-{
-    fn one() -> Self {
-        CheckedInt::new(T::one())
-    }
-}
-
-impl<T> ZRingValue for CheckedInt<T>
-where
-    T: Ord
-        + ::num::traits::Zero
-        + ::num::traits::One
-        + CheckedAdd
-        + CheckedMul
-        + CheckedSub
-        + CheckedNeg
-        + Clone
-        + 'static,
-{
-    fn ge0(&self) -> bool {
-        self.value >= T::zero()
-    }
-}
-
-// Note: this should be generic in T, but the Rust compiler does not like it
-// complaining that it conflicts with some implementation in core.
-impl<T> From<CheckedInt<T>> for i64
-where
-    T: Into<i64>,
-{
-    fn from(value: CheckedInt<T>) -> Self {
-        value.value.into()
-    }
-}
-
-impl<T> From<T> for CheckedInt<T> {
-    fn from(value: T) -> Self {
-        Self { value }
-    }
-}
-
-impl<T> Display for CheckedInt<T>
-where
-    T: Display,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), Error> {
-        self.value.fmt(f)
-    }
-}
-
-#[cfg(test)]
-pub type CheckedI64 = CheckedInt<i64>;
-
-#[cfg(test)]
-mod checked_integer_ring_tests {
-    use super::*;
-
-    #[test]
-    fn fixed_integer_tests() {
-        assert_eq!(0i64, CheckedI64::zero().into());
-        assert_eq!(1i64, CheckedI64::one().into());
-        let two = CheckedI64::one().add_by_ref(&CheckedI64::one());
-        assert_eq!(2i64, two.clone().into());
-        assert_eq!(-2i64, two.neg_by_ref().into());
-        assert_eq!(-4i64, two.clone().mul_by_ref(&two.neg_by_ref()).into());
-        let mut three = two;
-        three.add_assign_by_ref(&CheckedI64::from(1i64));
-        assert_eq!(3i64, three.clone().into());
-        assert_eq!(false, three.is_zero());
-    }
-
-    #[test]
-    #[should_panic]
-    fn overflow_test() {
-        let max = CheckedI64::from(i64::MAX);
-        let _ = max.add_by_ref(&CheckedI64::one());
     }
 }
